+++ src/widgets/AverageVisitors.php
@@ -77,7 +77,7 @@ class AverageVisitors extends Widget
      */
     public static function icon(): ?string
     {
-        return Craft::getAlias('@vnali/counter/icon-mask.svg');
+        return Craft::getAlias('@vnali/counter/icon-mask-widget.svg');
     }
 
     /**