+++ src/widgets/DecliningPages.php
@@ -66,7 +66,7 @@ class DecliningPages extends Widget
      */
     public static function icon(): ?string
     {
-        return Craft::getAlias('@vnali/counter/icon-mask.svg');
+        return Craft::getAlias('@vnali/counter/icon-mask-widget.svg');
     }
 
     /**