+++ src/widgets/MaxOnline.php
@@ -81,7 +81,7 @@ class MaxOnline extends Widget
      */
     public static function icon(): ?string
     {
-        return Craft::getAlias('@vnali/counter/icon-mask.svg');
+        return Craft::getAlias('@vnali/counter/icon-mask-widget.svg');
     }
 
     /**