+++ src/widgets/NextVisitedPages.php
@@ -74,7 +74,7 @@ class NextVisitedPages extends Widget
      */
     public static function icon(): ?string
     {
-        return Craft::getAlias('@vnali/counter/icon-mask.svg');
+        return Craft::getAlias('@vnali/counter/icon-mask-widget.svg');
     }