+++ src/widgets/NotVisitedPages.php
@@ -67,7 +67,7 @@ class NotVisitedPages extends Widget
      */
     public static function icon(): ?string
     {
-        return Craft::getAlias('@vnali/counter/icon-mask.svg');
+        return Craft::getAlias('@vnali/counter/icon-mask-widget.svg');
     }
 
     /**