+++ src/widgets/PageStatistics.php
@@ -64,7 +64,7 @@ class PageStatistics extends Widget
      */
     public static function icon(): ?string
     {
-        return Craft::getAlias('@vnali/counter/icon-mask.svg');
+        return Craft::getAlias('@vnali/counter/icon-mask-widget.svg');
     }