+++ src/widgets/VisitsRecent.php
@@ -87,7 +87,7 @@ class VisitsRecent extends Widget
      */
     public static function icon(): ?string
     {
-        return Craft::getAlias('@vnali/counter/icon-mask.svg');
+        return Craft::getAlias('@vnali/counter/icon-mask-widget.svg');
     }
 
     /**